backend/requirements.txt
# Sealionyx Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Cryptography
cryptography>=42.0.0

# Image handling (embedded seals)
Pillow>=10.0.0
piexif>=1.1.3

# Supabase
supabase>=2.3.0

# JWT & Auth
PyJWT>=2.8.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Development
black>=23.12.0
isort>=5.13.0
